model/requirements.txt
# 基础依赖
numpy>=1.20.0
torch>=1.8.0
torchaudio>=0.8.0
scipy>=1.6.0
librosa>=0.8.0
soundfile>=0.10.3

# Web服务相关
flask>=2.0.0
flask-cors>=3.0.10
requests>=2.25.0

# 配置和工具
pyyaml>=6.0
python-dotenv>=0.19.0
tqdm>=4.62.0  # 用于显示下载进度

# 音频处理
sounddevice>=0.4.4
matplotlib>=3.4.0

# 可选依赖
# tensorboard>=2.6.0  # 用于训练可视化
# nltk>=3.6.0  # 用于文本处理 